Switch to "offline boards"
This commit enables editing/selecting boards in offline mode.
Essentially, a single fetch is made to the database on load, adding all
the boards to an app-wide list. Subsequent board creations are then
added to this list, with a separate api post saving the board to the
database. This creates a seamless online and offline experience.

This commit also adds three builtin boards to choose from, hardcoded in
the initial app state.